pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "tasktriage"
version = "0.2.1"
description = "Analyzes handwritten task notes from a note-taking device using Claude AI to generate actionable GTD-based execution plans."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "langchain-anthropic>=0.3.0",
    "langchain-core>=0.3.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0",
    "google-api-python-client>=2.100.0",
    "google-auth>=2.23.0",
    "google-auth-oauthlib>=1.2.0",
    "cryptography>=42.0.0",
    "streamlit>=1.31.0",
    "pdf2image>=1.16.0",
    "babel>=2.17.0"
]

[project.scripts]
task-triage = "tasktriage:main"
tasktriage = "tasktriage:main"

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-cov>=4.0",
]
